+++ app/models/check_filter_value.rb
@@ -11,7 +11,7 @@ class CheckFilterValue < ApplicationRecord
   belongs_to :project
   belongs_to :user
   belongs_to :check
-  belongs_to :check_filter, touch: true
+  belongs_to :check_filter
 
   # Methods
 end