+++ app/models/sheet.rb
@@ -258,21 +258,6 @@ class Sheet < ApplicationRecord
     retry
   end
 
-  # TODO: Launch after subject "update", sheet "update", check "update"...?
-  # - [X] After Sheet Create
-  # - [X] After Sheet Update
-  # - [X] After Sheet Move
-  # - [X] After Another Sheet Update
-  # - [X] After Subject Randomized
-  # - [X] After Subject Unrandomized
-  # - [X] After Check Update (Check filter/check filter value) ?
-  def reset_checks!
-    project.checks.runnable.find_each do |check|
-      status_checks.where(check_id: check.id).first_or_create
-    end
-    status_checks.update_all(failed: nil)
-  end
-
   def create_notifications!
     return unless design.notifications_enabled?
     users.each do |u|