+++ db/migrate/20180606154604_add_last_run_at_to_checks.rb
@@ -1,0 +1,6 @@
+class AddLastRunAtToChecks < ActiveRecord::Migration[5.2]
+  def change
+    add_column :checks, :last_run_at, :datetime
+    add_index :checks, :last_run_at
+  end
+end